# server/pkg/models/absolute_discount.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * CHECK24 GenDev 7 API
 *
 * API for the 7th CHECK24 GenDev challenge providing product offerings from five different internet providers
 *
 * API version: dev
 */

package models

import (
	"errors"
)

type AbsoluteDiscount struct {
	ValueInCent int32 `json:"valueInCent"`

	MinOrderValueInCent *int32 `json:"minOrderValueInCent,omitempty"`
}

// AssertAbsoluteDiscountRequired checks if the required fields are not zero-ed
func AssertAbsoluteDiscountRequired(obj AbsoluteDiscount) error {
	elements := map[string]interface{}{
		"valueInCent": obj.ValueInCent,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertAbsoluteDiscountConstraints checks if the values respects the defined constraints
func AssertAbsoluteDiscountConstraints(obj AbsoluteDiscount) error {
	if obj.ValueInCent < 0 {
		return &ParsingError{Param: "ValueInCent", Err: errors.New(ErrMsgMinValueConstraint)}
	}
	if obj.MinOrderValueInCent != nil && *obj.MinOrderValueInCent < 0 {
		return &ParsingError{Param: "MinOrderValueInCent", Err: errors.New(ErrMsgMinValueConstraint)}
	}
	return nil
}