# server/pkg/models/address.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * CHECK24 GenDev 7 API
 *
 * API for the 7th CHECK24 GenDev challenge providing product offerings from five different internet providers
 *
 * API version: dev
 */

package models

type Address struct {
	Street string `json:"street"`

	HouseNumber string `json:"houseNumber,omitempty"`

	City string `json:"city"`

	PostalCode string `json:"postalCode"`

	CountryCode CountryCode `json:"countryCode"`
}

// AssertAddressRequired checks if the required fields are not zero-ed
func AssertAddressRequired(obj Address) error {
	elements := map[string]interface{}{
		"street":      obj.Street,
		"city":        obj.City,
		"postalCode":  obj.PostalCode,
		"countryCode": obj.CountryCode,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertAddressConstraints checks if the values respects the defined constraints
func AssertAddressConstraints(obj Address) error {
	return nil
}