# server/pkg/models/internet_products_cursor.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * CHECK24 GenDev 7 API
 *
 * API for the 7th CHECK24 GenDev challenge providing product offerings from five different internet providers
 *
 * API version: dev
 */

package models

// InternetProductsCursor - Response containing a list of internet products and version info
type InternetProductsCursor struct {
	Version string `json:"version,omitempty"`

	// Cursor to retrieve the first batch of products
	NextCursor string `json:"nextCursor,omitempty"`
}

// AssertInternetProductsCursorRequired checks if the required fields are not zero-ed
func AssertInternetProductsCursorRequired(obj InternetProductsCursor) error {
	return nil
}

// AssertInternetProductsCursorConstraints checks if the values respects the defined constraints
func AssertInternetProductsCursorConstraints(obj InternetProductsCursor) error {
	return nil
}