# server/pkg/models/internet_products_response.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * CHECK24 GenDev 7 API
 *
 * API for the 7th CHECK24 GenDev challenge providing product offerings from five different internet providers
 *
 * API version: dev
 */

package models

import (
	"encoding"
	"encoding/json"
)

const INTERNET_PRODUCTS_RESPONSE_VERSION = "1.0.0"

// InternetProductsResponse - Response containing a list of internet products and version info
type InternetProductsResponse struct {
	Products []InternetProduct `json:"products,omitempty"`

	// Cursor to retrieve the next batch of products, or null if finished
	NextCursor string `json:"nextCursor,omitempty"`
}

// AssertInternetProductsResponseRequired checks if the required fields are not zero-ed
func AssertInternetProductsResponseRequired(obj InternetProductsResponse) error {
	for _, el := range obj.Products {
		if err := AssertInternetProductRequired(el); err != nil {
			return err
		}
	}
	return nil
}

// AssertInternetProductsResponseConstraints checks if the values respects the defined constraints
func AssertInternetProductsResponseConstraints(obj InternetProductsResponse) error {
	for _, el := range obj.Products {
		if err := AssertInternetProductConstraints(el); err != nil {
			return err
		}
	}
	return nil
}

// Implementation of the BinaryMarshaller interface
func (obj InternetProductsResponse) MarshalBinary() ([]byte, error) {
	return json.Marshal(obj)
}

// Implementation of the BinaryUnmarshaller interface
func (obj *InternetProductsResponse) UnmarshalBinary(data []byte) error {
	return json.Unmarshal(data, obj)
}

// Ensure that InternetProduct implements the BinaryMarshaler and BinaryUnmarshaler interfaces
var _ encoding.BinaryMarshaler = (*InternetProductsResponse)(nil)
var _ encoding.BinaryUnmarshaler = (*InternetProductsResponse)(nil)