# server/pkg/models/product_info.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * CHECK24 GenDev 7 API
 *
 * API for the 7th CHECK24 GenDev challenge providing product offerings from five different internet providers
 *
 * API version: dev
 */

package models

import (
	"errors"
)

type ProductInfo struct {

	// in Mbps
	Speed int32 `json:"speed"`

	ConnectionType ConnectionType `json:"connectionType"`

	Tv *string `json:"tv,omitempty"`

	UnthrottledCapacityMb *int32 `json:"unthrottledCapacityMb,omitempty"`
}

// AssertProductInfoRequired checks if the required fields are not zero-ed
func AssertProductInfoRequired(obj ProductInfo) error {
	elements := map[string]interface{}{
		"speed":          obj.Speed,
		"connectionType": obj.ConnectionType,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertProductInfoConstraints checks if the values respects the defined constraints
func AssertProductInfoConstraints(obj ProductInfo) error {
	if obj.Speed < 0 {
		return &ParsingError{Param: "Speed", Err: errors.New(ErrMsgMinValueConstraint)}
	}
	if obj.UnthrottledCapacityMb != nil && *obj.UnthrottledCapacityMb < 0 {
		return &ParsingError{Param: "UnthrottledCapacityMb", Err: errors.New(ErrMsgMinValueConstraint)}
	}
	return nil
}

func CanonicalizeProductInfo(obj ProductInfo) ProductInfo {
	if obj.Tv != nil && *obj.Tv == "" {
		obj.Tv = nil
	}
	if obj.UnthrottledCapacityMb != nil && *obj.UnthrottledCapacityMb == 0 {
		obj.UnthrottledCapacityMb = nil
	}
	return obj
}